+++ toolchain/scripts/modules/assert/project/name.ps1
@@ -13,4 +13,6 @@ function Assert-ProjectName {
   if ($Name -like "*}*") { Write-Host "error : char } is not allowed in project name" -ForegroundColor Red; exit 1 }
   if ($Name -like "*#*") { Write-Host "error : char # is not allowed in project name" -ForegroundColor Red; exit 1 }
   if ($Name.Length -gt 16) { Write-Host "error : project name is too long (16 char max)" -ForegroundColor Red; exit 1 }
+
+  Write-Host "project name is compliant" -ForegroundColor Green
 }