+++ toolchain/scripts/modules/assert/raine/config.ps1
@@ -1,0 +1,19 @@
+function Assert-RaineConfig {
+  Write-Host "Asserting Raine config" -ForegroundColor Yellow
+  if (-Not(Resolve-Path -Path "$($Config.project.buildPath)\raine\config\raine32_sdl.cfg")) {
+    Write-Host "$($Config.project.buildPath)\raine\config\raine32_sdl.cfg not found" -ForegroundColor Red
+    exit 1
+  }
+  if (-Not(Resolve-Path -Path "$($Config.project.buildPath)\raine\config\default.cfg")) {
+    Write-Host "$($Config.project.buildPath)\raine\config\default.cfg not found" -ForegroundColor Red
+    exit 1
+  }
+  if (-Not(Resolve-Path -Path "$($Config.project.buildPath)\raine\config\fullscreen.cfg")) {
+    Write-Host "$($Config.project.buildPath)\raine\config\fullscreen.cfg not found" -ForegroundColor Red
+    exit 1
+  }
+  if (-Not(Resolve-Path -Path "$($Config.project.buildPath)\raine\config\yuv.cfg")) {
+    Write-Host "$($Config.project.buildPath)\raine\config\yuv.cfg not found" -ForegroundColor Red
+    exit 1
+  }
+}