+++ toolchain/scripts/modules/resolve/template-path.ps1
@@ -1,0 +1,9 @@
+function Resolve-TemplatePath {
+  param(
+    [Parameter(Mandatory=$true)][String] $Path
+  )
+
+  $resolvedPath = Get-TemplatePath -Path $Path
+
+  return (Resolve-Path -Path $resolvedPath).Path
+}