+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+## 2.7.0
+
+  - add raine config switcher
+  - externalize dependencies into manifest.xml
+  - add mak clean:build
+  - improve toolchain scripts
+  - add mak.log in build folder
+  - add gcc.log in build\project.name folder
+
 ## 2.6.1
 
   - update sh get headers script